Final file (potentially)
I added comments to the pipeline and added the forwarding stall for step 3. I also went through and checked our instruction count, and rn it's 91069 compared to Dan's 90080. This comes from the forwarding stall i added in step 3, us using 10 instead of 10-1, and for step 3 we also used 10 or 11 as our stall, whereas Dan used 8 and 9.